# ShopClient/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ProductTypes": {
      "get": {
        "tags": [
          "ProductTypes"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductType"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductType"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProductTypes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              }
            }
          }
        }
      }
    },
    "/api/ProductTypes/{id}": {
      "get": {
        "tags": [
          "ProductTypes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProductTypes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ProductTypes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Shops": {
      "get": {
        "tags": [
          "Shops"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/shopGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/shopGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/shopGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Shops"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/shopPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/shopPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/shopPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/shopGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/shopGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/shopGetDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Shops/{id}": {
      "get": {
        "tags": [
          "Shops"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/shopGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/shopGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/shopGetDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Shops"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/shopPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/shopPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/shopPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Shops"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TSs": {
      "get": {
        "tags": [
          "TSs"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TSGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TSGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TSGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TSs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TSPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TSPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TSPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TSGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TSGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TSGetDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/TSs/{id}": {
      "get": {
        "tags": [
          "TSs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TSGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TSGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TSGetDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TSs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TSPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TSPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TSPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TSs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProductType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TSGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "car_number": {
            "type": "integer",
            "format": "int32"
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "typeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TSPostDto": {
        "type": "object",
        "properties": {
          "car_number": {
            "type": "integer",
            "format": "int32"
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "typeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "shopGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "courierId": {
            "type": "integer",
            "format": "int32"
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "date_time_order": {
            "type": "string",
            "format": "date-time"
          },
          "date_time_delivery": {
            "type": "string",
            "format": "date-time"
          },
          "date_time_delivery_actual": {
            "type": "string",
            "format": "date-time"
          },
          "order_number": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "typeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "shopPostDto": {
        "type": "object",
        "properties": {
          "courierId": {
            "type": "integer",
            "format": "int32"
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "date_time_order": {
            "type": "string",
            "format": "date-time"
          },
          "date_time_delivery": {
            "type": "string",
            "format": "date-time"
          },
          "date_time_delivery_actual": {
            "type": "string",
            "format": "date-time"
          },
          "order_number": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "typeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}