# server/Dto/TSGetDto.cs
namespace server.Dto;

public class TSGetDto
{
    public int Id { get; set; }
    public int car_number { get; set; }
    public string model { get; set; } = string.Empty;
    public int typeId { get; set; }
}